--- src/db/models/document.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(sqlx::FromRow, Deserialize, Serialize)]
-/// Model for documents.
-pub struct Document {
-    /// Unique document identifier.
-    pub doc_id: String,
-}
