--- src/db/models/library.rs
@@ -1,8 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(sqlx::FromRow, Deserialize, Serialize)]
-/// Model for library (collection).
-pub struct Library {
-    /// Materialized path to the library
-    pub mpath: String,
-}
