+++ src/db/models/publication_has_publication_versions/mod.rs
@@ -1,4 +1,17 @@
+use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
+
+pub mod manager;
+
+/// Trait for managing transactions on publication has publication versions.
+#[async_trait]
+pub trait TxManager {
+    /// Upsert a bulk of `publication_has_publication_versions` into the database.
+    async fn insert_bulk(
+        &mut self,
+        publication_has_publication_versions: Vec<PublicationHasPublicationVersions>,
+    ) -> anyhow::Result<()>;
+}
 
 #[derive(sqlx::FromRow, Deserialize, Serialize, Clone, Debug)]
 /// Model for publication which contain publication versions.
