--- src/db/statements/mod.rs
@@ -1,4 +1,0 @@
-// Contains all the statements for the database
-pub mod inserts;
-// Contains all the queries for the database
-pub mod queries;
